Move expressions ASTs to Expressions directory
