+++ src/Ast/Expressions/Expression.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Serhii\GoodbyeHtml\Ast\Expressions;
+
+use Serhii\GoodbyeHtml\Ast\Node;
+
+interface Expression extends Node
+{
+}
