+++ src/Ast/Expressions/InfixExpression.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml\Ast;
+namespace Serhii\GoodbyeHtml\Ast\Expressions;
 
 use Serhii\GoodbyeHtml\Token\Token;
