+++ src/Ast/Statements/IfStatement.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Serhii\GoodbyeHtml\Ast\Statements;
 
-use Serhii\GoodbyeHtml\Ast\Expression;
+use Serhii\GoodbyeHtml\Ast\Expressions\Expression;
 use Serhii\GoodbyeHtml\Token\Token;
 
 readonly class IfStatement implements Statement
