+++ src/Evaluator/Evaluator.php
@@ -10,18 +10,18 @@ use Serhii\GoodbyeHtml\Ast\Statements\ExpressionStatement;
 use Serhii\GoodbyeHtml\Ast\Literals\FloatLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\HtmlStatement;
 use Serhii\GoodbyeHtml\Ast\Statements\IfStatement;
-use Serhii\GoodbyeHtml\Ast\InfixExpression;
+use Serhii\GoodbyeHtml\Ast\Expressions\InfixExpression;
 use Serhii\GoodbyeHtml\Ast\Literals\IntegerLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\LoopStatement;
 use Serhii\GoodbyeHtml\Obj\Env;
 use Serhii\GoodbyeHtml\Obj\Obj;
 use Serhii\GoodbyeHtml\Ast\Node;
 use Serhii\GoodbyeHtml\Ast\Literals\NullLiteral;
-use Serhii\GoodbyeHtml\Ast\PrefixExpression;
+use Serhii\GoodbyeHtml\Ast\Expressions\PrefixExpression;
 use Serhii\GoodbyeHtml\Ast\Statements\Program;
 use Serhii\GoodbyeHtml\Ast\Literals\StringLiteral;
-use Serhii\GoodbyeHtml\Ast\TernaryExpression;
-use Serhii\GoodbyeHtml\Ast\VariableExpression;
+use Serhii\GoodbyeHtml\Ast\Expressions\TernaryExpression;
+use Serhii\GoodbyeHtml\Ast\Expressions\VariableExpression;
 use Serhii\GoodbyeHtml\Obj\BlockObj;
 use Serhii\GoodbyeHtml\Obj\BooleanObj;
 use Serhii\GoodbyeHtml\Obj\ErrorObj;
